~ | Formatting the entire codebase to linux kernel coding style

+++ lib/lib.h
@@ -13,11 +13,12 @@ int	strlenk(const char *str);
 int	writek(int c, int len);
 int	printk(const char *str, ...);
 int	strncmpk(const char *s1, const char *s2, unsigned int n);
-void	*memsetk(void *ptr, int value, unsigned long num);
+void	*memsetk(void *ptr, i32 value, u32 num);
 int	strcmpk(const char *s1, const char *s2);
-bool	isprintk(int c);
+bool	isprintk(i32 c);
 
-static inline bool is_str_equal(const char *s1, const char *s2) {
+static inline bool is_str_equal(const char *s1, const char *s2)
+{
 	return(0 == strcmpk(s1, s2));
 }
